v44-00

+++ cloonix/common/lib_io_clownix/src/util_sock.c
@@ -34,6 +34,7 @@
 
 #include "io_clownix.h"
 #include "util_sock.h"
+#include "channel.h"
 
 
 /*****************************************************************************/
@@ -91,6 +92,27 @@ void util_free_fd(int fd)
 /*---------------------------------------------------------------------------*/
 
 /*****************************************************************************/
+/*
+int util_read_brakes_on(char *ptr, int max, int fd, int llid)
+{
+  int rx_len;
+  rx_len = read (fd, ptr, max);
+  if (rx_len < 0)
+    {
+    if ((errno == EAGAIN) || (errno ==EINTR))
+      {
+      rx_len = 0;
+      channel_rx_slow_down(llid);
+      }
+    }
+  else if (rx_len == 0)
+    rx_len = -1;
+  return rx_len;
+}
+*/
+/*---------------------------------------------------------------------------*/
+
+/*****************************************************************************/
 int util_read (char *ptr, int max, int fd)
 {
   int rx_len;
@@ -114,43 +136,43 @@ static void tempo_connect(int no_use)
 /*---------------------------------------------------------------------------*/
 
 /*****************************************************************************/
-static int test_file_is_socket(char *path)
-{
-  int result = -1;
-  struct stat stat_file;
-  if (!stat(path, &stat_file))
-    {
-    if (S_ISSOCK(stat_file.st_mode))
-      result = 0;
-    }
-  return result;
-}
-/*---------------------------------------------------------------------------*/
-
-/*****************************************************************************/
 int util_nonblock_client_socket_unix(char *pname, int *fd)
 {
   int len,  sock, result = -1;
   struct sockaddr_un addr;
-  if (!test_file_is_socket(pname))
+  struct stat stat_file;
+  signal(SIGPIPE, SIG_IGN);
+  if ((!pname) || (strlen(pname) == 0))
     {
-    sock = socket (AF_UNIX,SOCK_STREAM,0);
-    if (sock <= 0)
-      KOUT(" ");
-    memset (&addr, 0, sizeof (struct sockaddr_un));
-    addr.sun_family = AF_UNIX;
-    strcpy(addr.sun_path, pname);
-    len = sizeof (struct sockaddr_un);
-    nonblock_fd(sock);
-    result = connect(sock,(struct sockaddr *) &addr, len);
-    if (result != 0)
-      {
-      close(sock);
-      KERR("%s %d", pname, errno);
-      }
-    else
-      *fd = sock;
+    KERR("ERROR");
+    return -1;
     }
+  if (stat(pname, &stat_file))
+    {
+    KERR("ERROR %s", pname);
+    return -1;
+    }
+  if (!S_ISSOCK(stat_file.st_mode))
+    {
+    KERR("ERROR %s", pname);
+    return -1;
+    }
+  sock = socket (AF_UNIX,SOCK_STREAM,0);
+  if (sock <= 0)
+    KOUT("ERROR");
+  memset (&addr, 0, sizeof (struct sockaddr_un));
+  addr.sun_family = AF_UNIX;
+  strcpy(addr.sun_path, pname);
+  len = sizeof (struct sockaddr_un);
+  nonblock_fd(sock);
+  result = connect(sock,(struct sockaddr *) &addr, len);
+  if (result != 0)
+    {
+    close(sock);
+    KERR("ERROR %s %d", pname, errno);
+    }
+  else
+   *fd = sock;
   return result;
 }
 /*---------------------------------------------------------------------------*/
@@ -160,6 +182,9 @@ int util_nonblock_client_socket_inet(uint32_t ip, uint16_t port)
 {
   int sock, result;
   struct sockaddr_in adresse;
+  char ip_string[MAX_NAME_LEN];
+  int_to_ip_string (ip, ip_string);
+  signal(SIGPIPE, SIG_IGN);
   sock = socket (AF_INET,SOCK_STREAM,0);
   if (sock < 0)
     KOUT(" ");
@@ -181,9 +206,14 @@ int util_nonblock_client_socket_inet(uint32_t ip, uint16_t port)
       }
     else
       {
-      KERR("ERROR %X %hu %d", ip, port, errno);
+      if (errno == ENETUNREACH)
+        {
+        KERR("ERROR %s %hu Network is unreachable", ip_string, port);
+        }
+      else
+        KERR("ERROR %s %hu %d", ip_string, port, errno);
       if (close(sock))
-        KERR("ERROR %X %hu %d", ip, port, errno);
+        KERR("ERROR %s %hu %d", ip_string, port, errno);
       result = -1;
       }
     }
@@ -198,6 +228,12 @@ int util_client_abstract_socket_unix(char *pname, int *fd)
   struct sockaddr_un addr;
   struct sigaction act;
   struct sigaction oldact;
+  signal(SIGPIPE, SIG_IGN);
+  if ((!pname) || (strlen(pname) == 0))
+    {
+    KERR("ERROR");
+    return -1;
+    }
   sock = socket (AF_UNIX,SOCK_STREAM,0);
   if (sock <= 0)
     KOUT("ERROR");
@@ -238,38 +274,53 @@ int util_client_socket_unix(char *pname, int *fd)
   struct sockaddr_un addr;
   struct sigaction act;
   struct sigaction oldact;
-  if (!test_file_is_socket(pname))
+  struct stat stat_file;
+  signal(SIGPIPE, SIG_IGN);
+  if ((!pname) || (strlen(pname) == 0))
     {
-    sock = socket (AF_UNIX,SOCK_STREAM,0);
-    if (sock <= 0)
-      KOUT(" ");
-    memset (&addr, 0, sizeof (struct sockaddr_un));
-    addr.sun_family = AF_UNIX;
-    strcpy(addr.sun_path, pname);
-    len = sizeof (struct sockaddr_un);
+    KERR("ERROR");
+    return -1;
+    }
+  if (stat(pname, &stat_file))
+    {
+    KERR("ERROR %s", pname);
+    return -1;
+    }
+  if (!S_ISSOCK(stat_file.st_mode))
+    {
+    KERR("ERROR %s", pname);
+    return -1;
+    }
+  sock = socket (AF_UNIX,SOCK_STREAM,0);
+  if (sock <= 0)
+    KOUT(" ");
+  memset (&addr, 0, sizeof (struct sockaddr_un));
+  addr.sun_family = AF_UNIX;
+  strcpy(addr.sun_path, pname);
+  len = sizeof (struct sockaddr_un);
 
-    memset(&act, 0, sizeof(act));
-    sigfillset(&act.sa_mask);
-    act.sa_flags = 0;
-    act.sa_handler = tempo_connect;
-    sigaction(SIGALRM, &act, &oldact);
+  memset(&act, 0, sizeof(act));
+  sigfillset(&act.sa_mask);
+  act.sa_flags = 0;
+  act.sa_handler = tempo_connect;
+  sigaction(SIGALRM, &act, &oldact);
 
-    alarm(5);
-    result = connect(sock,(struct sockaddr *) &addr, len);
+  alarm(5);
+  result = connect(sock,(struct sockaddr *) &addr, len);
 
-    sigaction(SIGALRM, &oldact, NULL);
-    alarm(0);
+  sigaction(SIGALRM, &oldact, NULL);
+  alarm(0);
 
-    if (result != 0)
-      {
-      close(sock);
-      printf("NO SERVER LISTENING TO %s\n", pname);
-      }
-    else
-      {
-      *fd = sock;
-      nonblock(*fd);
-      }
+  if (result != 0)
+    {
+    close(sock);
+    printf("NO SERVER LISTENING TO %s\n", pname);
+    KERR("ERROR NO SERVER LISTENING TO %s", pname);
+    }
+  else
+    {
+    *fd = sock;
+    nonblock(*fd);
     }
   return result;
 }
@@ -282,6 +333,12 @@ int util_proxy_client_socket_unix(char *pname, int *fd)
   struct sockaddr_un addr;
   struct sigaction act;
   struct sigaction oldact;
+  signal(SIGPIPE, SIG_IGN);
+  if ((!pname) || (strlen(pname) == 0))
+    {
+    KERR("ERROR");
+    return -1;
+    }
   sock = socket (AF_UNIX,SOCK_STREAM,0);
   if (sock <= 0)
     KOUT(" ");
@@ -323,6 +380,7 @@ int util_client_socket_inet(uint32_t ip, uint16_t port, int *fd)
   int   sock, result;
   struct sockaddr_in adresse;
   *fd = -1;
+  signal(SIGPIPE, SIG_IGN);
   sock = socket (AF_INET,SOCK_STREAM,0);
   if (sock <= 0)
     KOUT(" ");
@@ -364,6 +422,12 @@ int util_socket_listen_unix(char *pname)
   int ret, fd, len;
   struct sockaddr_un serv;
 
+  signal(SIGPIPE, SIG_IGN);
+  if ((!pname) || (strlen(pname) == 0))
+    {
+    KERR("ERROR");
+    return -1;
+    }
   if (strlen(pname) >= 108)
     KOUT("%d", (int)(strlen(pname)));
   unlink (pname);
@@ -400,7 +464,13 @@ int util_socket_unix_dgram(char *pname)
 { 
   int ret, fd, len;
   struct sockaddr_un serv;
+  signal(SIGPIPE, SIG_IGN);
   
+  if ((!pname) || (strlen(pname) == 0))
+    {
+    KERR("ERROR");
+    return -1;
+    }
   if (strlen(pname) >= 108)
     KOUT("%d", (int)(strlen(pname)));
   unlink (pname);
@@ -427,6 +497,7 @@ int util_socket_listen_inet( uint16_t port )
 {
   struct sockaddr s_addr;
   int fd, optval=1, nodelay=1;
+  signal(SIGPIPE, SIG_IGN);
   memset(&s_addr, 0, sizeof(s_addr));
   s_addr.sa_family = AF_INET;
   ((struct sockaddr_in*)&s_addr)->sin_addr.s_addr = htonl(INADDR_ANY);
